Create project infrastructure for KUnit in security Medusa - move existing tests from drivers to security Medusa - add option to run KUnit tests in build.sh - prepare basic Kconfigs and Makefile for running tests
