+++ .kunitconfig
@@ -1,3 +1,9 @@
 CONFIG_KUNIT=y
-CONFIG_MISC_EXAMPLE=y
-CONFIG_MISC_EXAMPLE_TEST=y
+CONFIG_MEDUSA_L3_TESTING=y
+
+CONFIG_NET=y
+CONFIG_SECURITY=y
+CONFIG_SECURITY_MEDUSA=y
+CONFIG_SECURITY_MEDUSA_TESTING=y
+CONFIG_CHECKPOINT_RESTORE=y
+CONFIG_SYSVIPC=y
