+++ drivers/misc/Makefile
@@ -57,5 +57,3 @@ obj-y				+= cardreader/
 obj-$(CONFIG_PVPANIC)   	+= pvpanic.o
 obj-$(CONFIG_HABANA_AI)		+= habanalabs/
 obj-$(CONFIG_XILINX_SDFEC)	+= xilinx_sdfec.o
-obj-$(CONFIG_MISC_EXAMPLE_TEST) += model-tests.o
-obj-$(CONFIG_MISC_EXAMPLE_TEST) += vsmodel-tests.o
