+++ include/linux/medusa/l3/med_model.h
@@ -2,7 +2,7 @@
 #define _MEDUSA_MODEL_H
 
 #include <linux/medusa/l3/config.h>
-#include <linux/medusa/l3/vsmodel.h>
+#include <linux/medusa/l3/vs_model.h>
 
 extern int medusa_authserver_magic;
