+++ security/medusa/Makefile
@@ -8,10 +8,11 @@ subdir-$(CONFIG_SECURITY_MEDUSA)	+= l0
 subdir-$(CONFIG_SECURITY_MEDUSA)	+= l1
 subdir-$(CONFIG_SECURITY_MEDUSA)        += l2
 subdir-$(CONFIG_SECURITY_MEDUSA)        += l3
+subdir-$(CONFIG_SECURITY_MEDUSA)        += testing
 
 #obj-$(CONFIG_SECURITY_MEDUSA) += l0/
 obj-$(CONFIG_SECURITY_MEDUSA) += l1/
 obj-$(CONFIG_SECURITY_MEDUSA) += l2/
 obj-$(CONFIG_SECURITY_MEDUSA) += l3/
 obj-$(CONFIG_SECURITY_MEDUSA_L4_CONSTABLE) += l4-constable/
-
+obj-$(CONFIG_SECURITY_MEDUSA_TESTING) += testing/
