+++ security/medusa/testing/Makefile
@@ -1,0 +1,13 @@
+#
+# Makefile for building the BASC module as part of the kernel tree.
+#
+
+#subdir-$(CONFIG_MEDUSA_L1_TESTING)	:= l1
+#subdir-$(CONFIG_MEDUSA_L2_TESTING)        += l2
+subdir-$(CONFIG_MEDUSA_L3_TESTING)        += l3
+#subdir-$(CONFIG_MEDUSA_L4_TESTING)        += l4
+
+#obj-$(CONFIG_MEDUSA_L1_TESTING) += l1/
+#obj-$(CONFIG_MEDUSA_L2_TESTING) += l2/
+obj-$(CONFIG_MEDUSA_L3_TESTING) += l3/
+#obj-$(CONFIG_MEDUSA_L4_TESTING) += l4/
