+++ security/medusa/testing/l3/Kconfig
@@ -1,0 +1,3 @@
+config MEDUSA_L3_TESTING
+	bool "Tests for L3 layer"
+	depends on KUNIT
