+++ security/medusa/testing/l3/Makefile
@@ -1,0 +1,8 @@
+#
+# Makefile for the Linux kernel part of Medusa DS9 Security System
+#
+
+# note that the order is important here!
+
+obj-$(CONFIG_MEDUSA_L3_TESTING) := vs_model-tests.o
+obj-$(CONFIG_MEDUSA_L3_TESTING) += med_model-tests.o
