+++ security/medusa/testing/l3/med_model-tests.c
@@ -1,5 +1,5 @@
 #include <kunit/test.h>
-#include <linux/medusa/l3/model.h>
+#include <linux/medusa/l3/med_model.h>
 #include <linux/medusa/l1/task.h>
 
 static void fake_med_object_init(struct medusa_object_s *med_object)
